feat: broken test that needs to be fixed
we will fix this by making it so that a number node will return next,
which will be itself, when a number is passed in.  that way the tree
remains unambiguous and it enforces that we are down the number tree.

this may also mean that we need to play a failed motion again from the
top if the previous motion was a number... dang that is hard